Parameterized reset-domain for ant CLI and UCD Plugin.
doRestore method in src/com/ibm/dcm/Soma.java hard coded reset-domain=true
in the generation of the SOMA command that is sent to DataPower.  This change
allows that value to be passed to the dcm ant command line tool and via
the UCD plugin.  In addition, the additional domains feature did not reset
the domains like the first domain was being reset, therefore I added the
reset-domain paramter to that line of code as well.  The hard coded value
of reset-domain was true, so the default value for this is also true as to
not break existing users.

This change is to allow the user to decide wether or not to reset-domain
during a doRestore.  Problems present themselves with exporting and importing
backups between equivalent devices (DataPowers behind a Load Balancer).
Domain backups do not export passwords, this is problematic when importing
the config to an equivalent domain on another DataPower as the reset-domain
causes the objects to be removed and reimported, causing the password in
the password map alias to be lost.  As a result, the object that references
the password map alias will transition to a down state, potentially causing
a service outage.